fix window config bug
