+++ Assets/MyAssets/ViveWindowSize.asset
@@ -8,17 +8,17 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: Vive Window Size Config
-  m_EditorClassIdentifier: Assembly-CSharp::WindowSizeConfig
+  m_Script: {fileID: 11500000, guid: bb93d000ba601ec44ba34d5ab7605fce, type: 3}
+  m_Name: ViveWindowSize
+  m_EditorClassIdentifier: 
   HmdType: 0
   RadiusList:
   - Name: 0
     Inner: 0.2
-    Outer: 0.5
+    Outer: 0.4
   - Name: 1
     Inner: 0.3
-    Outer: 0.6
+    Outer: 0.5
   - Name: 2
     Inner: 0.4
-    Outer: 0.7
+    Outer: 0.6
