+++ Assets/MyAssets/ViveWindowSize.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe7077ff63c347f4181799c294d3eeed
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
