+++ Assets/Script/ConfigData.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 6c738b413468bc5408831da4792d0a80
-NativeFormatImporter:
+guid: 98352f51ed95d7e4ea9d6794b88ad77a
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
