+++ Assets/Script/ConfigData/ConditionData.cs.meta
