+++ Assets/Script/ConfigData/WindowSizeConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb93d000ba601ec44ba34d5ab7605fce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
